refactor: extract shared db abstractions
