+++ build.sbt
@@ -111,13 +111,20 @@ lazy val commonSettings = Seq(
 )
 
 lazy val root = (project in file("."))
-  .aggregate(core, fs, s3, tika, metrics, pg, docs)
+  .aggregate(core, db, fs, s3, tika, metrics, pg, docs)
   .settings(name := "graviton")
 
 lazy val core = project
   .in(file("modules/core"))
   .settings(
     name := "graviton-core"
+  )
+  .settings(commonSettings)
+
+lazy val db = project
+  .in(file("modules/db"))
+  .settings(
+    name := "graviton-db"
   )
   .settings(commonSettings)
 
@@ -163,7 +170,7 @@ lazy val dbcodegen = project
 
 lazy val pg = project
   .in(file("modules/pg"))
-  .dependsOn(core, dbcodegen)
+  .dependsOn(core, db, dbcodegen)
   .settings(
     name := "graviton-pg",
     // on-demand schema generation snapshot directory (checked into VCS)
@@ -282,7 +289,7 @@ lazy val docs = project
     projectName                                := "graviton",
     mainModuleName                             := (core / moduleName).value,
     projectStage                               := ProjectStage.ProductionReady,
-    ScalaUnidoc / unidoc / unidocProjectFilter := inProjects(core, fs, s3, tika, metrics, pg),
+    ScalaUnidoc / unidoc / unidocProjectFilter := inProjects(core, db, fs, s3, tika, metrics, pg),
     mdocIn                                     := baseDirectory.value / "src/main/mdoc",
     mdocOut                                    := baseDirectory.value / "target/mdoc",
     mdocVariables                              := Map("VERSION" -> version.value),
